refactor: NetworkModule에서 Service 관련 로직 분리
